[REF] Separate python objects in new files
