+++ product_usability/models/__init__.py
@@ -1,4 +1,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import product
+from . import product_template
+from . import product_supplierinfo
+from . import product_price_history
 from . import pricelist
