+++ product_usability/models/product_supplierinfo.py
@@ -1,0 +1,13 @@
+# © 2015-2016 Akretion (http://www.akretion.com)
+# @author Alexis de Lattre <alexis.delattre@akretion.com>
+# @author Raphaël Valyi <rvalyi@akretion.com>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import models, fields
+
+
+class ProductSupplierinfo(models.Model):
+    _inherit = 'product.supplierinfo'
+
+    name = fields.Many2one(
+        domain=[('supplier', '=', True), ('parent_id', '=', False)])
